# apifox/openapi.yaml
openapi: 3.0.3
info:
  title: dateOrFriends API
  version: 0.0.1-SNAPSHOT
  description: |
    dateOrFriends 后端接口文档（可直接导入 Apifox：导入 -> OpenAPI）。

    统一返回结构：{ code, message, data }
servers:
  - url: http://localhost:8888
security:
  - bearerAuth: []
...
paths:
...
  /admin/notice/{id}:
    get:
      tags: [Admin]
      summary: 公告管理-详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseNotice'
        '400':
          description: 参数不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
...
  /user/notices:
    get:
      tags: [User]
      summary: 用户端-公告列表
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageUserNotice'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /user/notices/popup:
    get:
      tags: [User]
      summary: 用户端-重要公告弹窗
      description: 返回“未读的重要公告”1条；若没有则 data=null
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserNoticeNullable'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /user/notices/{id}:
    get:
      tags: [User]
      summary: 用户端-公告详情（并标记已读）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserNotice'
        '400':
          description: 参数不合法/公告不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /user/notices/read:
    post:
      tags: [User]
      summary: 用户端-标记公告已读
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNoticeReadRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /user/notifications:
    get:
      tags: [User]
      summary: 用户端-站内通知收件箱
      security:
        - bearerAuth: []
      parameters:
        - name: read
          in: query
          required: false
          schema:
            type: integer
            format: int32
            description: 0=未读；1=已读；不传=全部
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageUserInboxNotification'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /user/notifications/unread-count:
    get:
      tags: [User]
      summary: 用户端-站内通知未读数量
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /user/notifications/read:
    post:
      tags: [User]
      summary: 用户端-标记站内通知已读
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNotificationReadRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '401':
          description: 未登录或令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponseVoid:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          description: 0=成功；非0=失败
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
...
    ApiResponseError:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          description: 业务错误码（例如 1000 参数不合法、1001 未授权、1002 无权限、2000 账号不存在 等）
          example: 1001
        message:
          type: string
          description: 错误信息
          example: 未登录或令牌无效
        data:
          nullable: true
...
    Notice:
      type: object
      required: [id, title, content, level, status, pinned, createdBy, updatedBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        level:
          type: string
          enum: [NORMAL, IMPORTANT]
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OFFLINE]
        pinned:
          type: boolean
        publishAt:
          type: string
          format: date-time
          nullable: true
        expireAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: integer
          format: int64
        updatedBy:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiResponseNotice:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/Notice'
...
    UserNotice:
      type: object
      required: [id, title, content, level, pinned, read]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        level:
          type: string
        pinned:
          type: boolean
        publishAt:
          type: string
          format: date-time
          nullable: true
        expireAt:
          type: string
          format: date-time
          nullable: true
        read:
          type: boolean
    PageUserNotice:
      type: object
      required: [total, page, size, items]
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserNotice'
    ApiResponseUserNotice:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/UserNotice'
    ApiResponseUserNoticeNullable:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/UserNotice'
          nullable: true
    ApiResponsePageUserNotice:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/PageUserNotice'
    MarkNoticeReadRequest:
      type: object
      required: [noticeId]
      properties:
        noticeId:
          type: integer
          format: int64
    ApiResponseLong:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          type: integer
          format: int64
...
    UserInboxNotification:
      type: object
      required: [id, notificationId, type, title, content, level, read, deliveredAt]
      properties:
        id:
          type: integer
          format: int64
        notificationId:
          type: integer
          format: int64
        type:
          type: string
        title:
          type: string
        content:
          type: string
        level:
          type: string
          enum: [NORMAL, IMPORTANT]
        linkUrl:
          type: string
          nullable: true
        payloadJson:
          type: string
          nullable: true
        expireAt:
          type: string
          format: date-time
          nullable: true
        notificationCreatedAt:
          type: string
          format: date-time
          nullable: true
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
    PageUserInboxNotification:
      type: object
      required: [total, page, size, items]
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserInboxNotification'
    ApiResponsePageUserInboxNotification:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/PageUserInboxNotification'
    MarkNotificationReadRequest:
      type: object
      required: [notificationId]
      properties:
        notificationId:
          type: integer
          format: int64